Update pyproject.toml. Removed  "mpl-image-compare",    # other dependencies moved it to optional-dependencies Changed mpl-image-compare to pytest-mpl
Pytest-mpl is a plugin for pytest that helps compare images of Matplotlib figures.

For each figure being tested, an image is created and compared to a reference image. If the difference (RMS) is greater than a set tolerance, the test fails. Another option is to hash the generated image and compare it to a known value.